requirements.txt
tensorflow-gpu
tensorflow_text
transformers